move provider resource parsing to config
