+++ dhctl/pkg/config/deckhouse_config.go
@@ -28,7 +28,6 @@ import (
 
 	"github.com/deckhouse/deckhouse/dhctl/pkg/app/options"
 	"github.com/deckhouse/deckhouse/dhctl/pkg/config/registry"
-	"github.com/deckhouse/deckhouse/dhctl/pkg/infrastructureprovider/providerdata"
 	"github.com/deckhouse/deckhouse/dhctl/pkg/log"
 	"github.com/deckhouse/deckhouse/dhctl/pkg/telemetry"
 )
@@ -77,7 +76,7 @@ func (c *DeckhouseInstaller) HasProviderModuleConfig() bool {
 	if c == nil || c.ProviderName == "" {
 		return false
 	}
-	target := providerdata.CloudProviderModuleName(c.ProviderName)
+	target := CloudProviderModuleName(c.ProviderName)
 	for _, mc := range c.ModuleConfigs {
 		if mc.Name == target {
 			return true
